+++ media_control/lib/src/client.dart
@@ -2,7 +2,7 @@ import 'package:http/http.dart'
     hide delete, get, head, patch, post, put, read, readBytes, runWithClient;
 import 'package:meta/meta.dart';
 
-import 'ua.dart' if (dart.library.html) 'ua_web.dart';
+import 'ua.dart' if (dart.library.js_interop) 'ua_web.dart';
 
 @internal
 final class OgHrefMediaClient extends BaseClient {
@@ -11,9 +11,8 @@ final class OgHrefMediaClient extends BaseClient {
   @override
   Future<StreamedResponse> send(BaseRequest request) {
     request
-      ..headers["user-agent"] =
-          requestUserAgent ?? "Mozilla/5.0 oghref/2 (Media classification)"
-      ..followRedirects = true;
+      ..headers["user-agent"] = requestUserAgent
+      ..followRedirects = false;
 
     return request.send();
   }