+++ media_control/lib/src/ua.dart
@@ -1,4 +1,4 @@
 import 'package:meta/meta.dart';
 
 @internal
-String? get requestUserAgent => null;
+String get requestUserAgent => "oghref_content_type_guard 1";